null transformer generate the null column and get the nan value
